requirements.txt
tensorflow-cpu==2.12.0
Flask==2.3.2
gunicorn==21.2.0
numpy==1.24.4
Pillow==10.0.1
Werkzeug==2.3.7